feat: add okx wallet and tests for gas calculation
