+++ package.json
@@ -62,7 +62,10 @@
     "prettier": "^2.8.8",
     "ts-jest": "^27.1.2",
     "turbo": "^1.10.16",
-    "typescript": "4.7.4"
+    "typescript": "4.7.4",
+    "jest-watch-typeahead": "^2.2.2",
+    "msw": "^1.3.2",
+    "whatwg-fetch": "^3.6.19"
   },
   "dependencies": {},
   "resolutions": {
