+++ packages/stores/src/__tests_e2e__/msw-server.ts
@@ -1,0 +1,4 @@
+// eslint-disable-next-line import/no-extraneous-dependencies
+import { setupServer } from "msw/node";
+
+export const server = setupServer();
