+++ packages/stores/src/currency-registrar/__tests__/unsafe-ibc.spec.ts
@@ -3,8 +3,8 @@ import { ChainInfoInner, ChainStore } from "@osmosis-labs/keplr-stores";
 import type { ChainInfo } from "@osmosis-labs/types";
 import { makeIBCMinimalDenom } from "@osmosis-labs/utils";
 
+import { mockChainInfos, mockIbcAssets } from "../../__tests_e2e__/mock-data";
 import { UnsafeIbcCurrencyRegistrar } from "../unsafe-ibc"; // make sure to import your class correctly
-import { mockChainInfos, mockIbcAssets } from "./mock-data";
 
 describe("UnsafeIbcCurrencyRegistrar", () => {
   let chainStoreMock: ChainStore;
