+++ packages/utils/package.json
@@ -17,11 +17,9 @@
     "sha.js": "^2.4.11"
   },
   "devDependencies": {
-    "jest-watch-typeahead": "^2.2.2",
     "jest-environment-jsdom": "^29.5.0",
     "jest": "^29.5.0",
-    "ts-jest": "^29.1.0",
-    "cross-fetch": "^4.0.0"
+    "ts-jest": "^29.1.0"
   },
   "lint-staged": {
     "*": [
