+++ packages/utils/setup-tests.ts
@@ -1,2 +1,2 @@
 // eslint-disable-next-line import/no-extraneous-dependencies
-import "cross-fetch/polyfill";
+import "whatwg-fetch";
