+++ packages/web/package.json
@@ -151,8 +151,6 @@
     "jest-environment-jsdom": "^29.5.0",
     "jest-in-case": "^1.0.2",
     "jest-launchdarkly-mock": "^2.1.0",
-    "jest-watch-typeahead": "^2.2.2",
-    "msw": "^1.3.2",
     "postcss": "^8.4.5",
     "prettier": "^2.8.7",
     "prettier-plugin-tailwindcss": "^0.2.0",
@@ -161,7 +159,6 @@
     "tailwindcss": "^3.2.4",
     "ts-jest": "^29.1.0",
     "tsconfig-paths": "^4.2.0",
-    "use-debugger-hooks": "^1.3.0",
-    "whatwg-fetch": "^3.6.19"
+    "use-debugger-hooks": "^1.3.0"
   }
 }
